Added in logic and tests to support only executing scripts once, but maintaining proper support for the callbacks and stuff. Requires some hackery in the 'preload' logic. Fixes #67
Essentially, don't actually preload a file if we've seen it before, and use the `onload` event of the first one to fire off the onload event of all that are similar. Then once it's loaded, go back to caching it since it'll be fast (and we don't have a callback to kick it off). Tests added to make sure this works. Quite a bit of code for a simple feature, but we made some room...